Move o portfolio para o index do usuário
