+++ app/pages/[username]/index.js
@@ -1,60 +1,137 @@
 import { useRouter } from 'next/router';
-import Head from 'next/head'
-import { Image, Text, Title, ThemeIcon } from '@mantine/core';
-import { getMonthsNames } from '@mantine/dates';
-import 'dayjs/locale/pt';
-import { FaCalendarCheck } from 'react-icons/fa';
 
-function convertDateToPT(dateString) {
-  const date = new Date(dateString);
-  return `${getMonthsNames('pt')[date.getMonth()]} de ${date.getFullYear()}`;
-}
+import Head from 'next/head';
 
+import {
+	Avatar,
+	Badge,
+	Button,
+	Card,
+	Container,
+	Divider,
+	Group,
+	Image,
+	Space,
+	SimpleGrid,
+	Text,
+	useMantineTheme,
+} from '@mantine/core';
+
+// TODO: Add user data to context
 async function fetchUser(username) {
-  const response = await fetch(process.env.STRAPI_API_URL + `/artic-users/?Username=${username}`);
-  if (response.ok) {
-    let users = await response.json();
-    if (users.length !== 0) {
-      return users[0];
-    }
-  }
-  return null;
+	const response = await fetch(process.env.STRAPI_API_URL + `/artic-users/?Username=${username}`);
+	if (response.ok) {
+		let users = await response.json();
+		if (users.length !== 0) {
+			return users[0];
+		}
+	}
+	return null;
 }
 
 export async function getServerSideProps({ params }) {
-  const user = await fetchUser(params.username);
-  return {
-    props: { user: user }
-  };
+	const user = await fetchUser(params.username);
+	return {
+		props: { user: user },
+	};
 }
 
-export default function User({ user }) {
-  const router = useRouter();
-  const username = router.query.username;
-  if (user === null) {
-    return (<>
-      <Head><title>Not Found</title></Head>
-      <Text>Usuário {username} não encontrado</Text>
-    </>);
-  }
-  return (<>
-    <Head>
-      <title>{user.Username}</title>
-    </Head>
-    {user.Picture !== null &&
-      <Image
-        src={process.env.STRAPI_API_URL + user.Picture.url}
-        width="300px"
-        alt="Foto de perfil do usuário."
-      />
-    }
-    {user.Name !== null &&
-      <Title order={2}>{user.Name}</Title>
-    }
-    <Text weight="bold" color="gray">@{user.Username}</Text>
-    {user.Bio !== null &&
-      <Text size="sm">{user.Bio}</Text>
-    }
-    <Text size="sm" color="gray"><FaCalendarCheck /> Ingressou em {convertDateToPT(user.created_at)}</Text>
-  </>);
+function BasicArtistInfo({ user }) {
+	const photo = user.Picture ? `${process.env.STRAPI_API_URL}${user.Picture.url}` : undefined;
+	return (
+		<Container size='md' padding='sm'>
+			<Card withBorder>
+				<Group>
+					<Card.Section>
+						<Avatar src={photo} alt='Foto do usuário' size='xl' />
+						{/* User profile pic */}
+					</Card.Section>
+					{user.Name && (
+						<Card.Section>
+							<Text weight={700} size='lg' style={{ lineHeight: 1.5 }}>
+								{user.Name}
+							</Text>
+						</Card.Section>
+					)}
+					{user.Bio && <Card.Section>{user.Bio}</Card.Section>}
+				</Group>
+			</Card>
+		</Container>
+	);
+}
+
+function ArtCard({ art }) {
+	const theme = useMantineTheme();
+
+	const secondaryColor =
+		theme.colorScheme === 'dark' ? theme.colors.dark[1] : theme.colors.gray[7];
+
+	const photo = art.media ? `${process.env.STRAPI_API_URL}${art.media.url}` : undefined;
+
+	return (
+		<Card shadow='md' withBorder>
+			<Card.Section>
+				<Image src={photo} height={200} alt='Imagem descritiva do projeto' />
+			</Card.Section>
+
+			<Group position='apart' style={{ marginBottom: 5, marginTop: theme.spacing.sm }}>
+				<Text weight={500}>{art.name}</Text>
+				<Group position='right' style={{ marginBottom: 5, marginTop: theme.spacing.sm }}>
+					{art.tags?.map((tag) => (
+						<Badge key={tag.id} color='pink' variant='light'>
+							{tag.name}
+						</Badge>
+					))}
+				</Group>
+			</Group>
+
+			{art.description && (
+				<Text size='sm' style={{ color: secondaryColor, lineHeight: 1.5 }}>
+					{art.description}
+				</Text>
+			)}
+
+			<Button variant='light' color='blue' fullWidth style={{ marginTop: 14 }}>
+				Ver mais
+			</Button>
+		</Card>
+	);
+}
+
+export default function WorksOfUser({ user }) {
+	const router = useRouter();
+	const { username } = router.query;
+
+	const { arts } = user;
+
+	return user ? (
+		<>
+			<Head>
+				<title>{username} | Obras</title>
+			</Head>
+			<BasicArtistInfo user={user} />
+			{arts && arts.length > 0 && (
+				<>
+					<Space h='xl' />
+					<Divider size='sm' />
+					<Space h='xl' />
+					<Space h='sm' />
+					<SimpleGrid cols={3} spacing='lg'>
+						{arts.map((art) => (
+							<Container key={art.id} m='sm' size='xs'>
+								<ArtCard art={art} />
+							</Container>
+						))}
+					</SimpleGrid>
+				</>
+			)}
+		</>
+	) : (
+		<>
+			<Head>
+				<title>Not Found</title>
+			</Head>
+			<Text>Usuário {username} não encontrado</Text>
+		</>
+	);
 }
